requirements.txt
streamlit
google-generativeai>=0.7.0
yfinance
pandas
pandas_ta
plotly